test: use templating with ytt to add support for windows and macos tests
Use `ytt` for generating templated workflows for testing on windows and
macos machines.
We use `ytt` here instead of the conventional manner of using a bash
script[^1] to spawn tests for different operating systems because we
also want to test in the context of a sandbox, which is supported only
on Linux, not on macOS and Windows.
This means that we need to conditionally include the `sandbox` plugin
while testing on Linux alone and rather than writing/finding (and then
using) another plugin for this or having multiple pipelines
(one that's linux-specific and one for windows and macos for each
workflow), it's much more maintainable to simply dynamically generate
and upload pipelines using `ytt`.

[^1]: https://github.com/JuliaCI/ytt-buildkite-plugin/blob/683479ac966db605eb6c5ae9f6297ce36fe9402c/.buildkite/pipeline.yml#L6-L30